refactor(web): keep member dialog trigger in caller
